core-backend/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.8.0

# ── WebSocket ─────────────────────────────────────────────────────────
websockets==12.0

# ── Redis (The Vault) ─────────────────────────────────────────────────
redis[asyncio]==5.0.3

# ── HTTP Client ───────────────────────────────────────────────────────
httpx==0.27.0

# ── Overlord Webhook dependencies ─────────────────────────────────────
aiohttp==3.9.3
aiofiles==23.2.1

# ── Browser Automation (QA Sniper) ────────────────────────────────────
playwright==1.43.0
# Sau khi build: RUN playwright install chromium --with-deps

# ── Scraping (Ghost Scraper Agent) ────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.1.0

# ── Scheduling ────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Utilities ─────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.9

# ── Machine Learning & Database ───────────────────────────────────────
asyncpg==0.29.0
scikit-learn==1.3.2
numpy==1.26.2
playwright-stealth==1.0.6

# ── Supabase SDK ───────────────────────────────────────────────────────
# https://github.com/supabase/supabase-py
supabase>=2.4.0

